requirements.txt
pdfplumber==0.11.0
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
openai==1.59.6
chromadb==0.6.3
networkx==3.4.2
matplotlib==3.9.4
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
streamlit==1.41.1
python-dotenv==1.0.1
